chore: moves Carousel to fs/components
