+++ packages/components/src/hooks/index.ts
@@ -2,3 +2,10 @@ export { default as UIProvider, Toast as ToastProps, useUI } from './UIProvider'
 export { useFadeEffect } from './useFadeEffect'
 export { useTrapFocus } from './useTrapFocus'
 export { useSearch } from './useSearch'
+export { useSlider } from './useSlider'
+export type {
+  UseSliderArgs,
+  SliderState,
+  SliderDispatch,
+  SlideDirection,
+} from './useSlider'
