+++ packages/components/src/molecules/Carousel/index.ts
@@ -1,0 +1,5 @@
+export { default } from './Carousel'
+export type { CarouselProps } from './Carousel'
+
+export { default as CarouselItem } from './CarouselItem'
+export type { CarouselItemProps } from './CarouselItem'
