--- packages/ui/src/components/molecules/Carousel/index.ts
@@ -1,2 +1,0 @@
-export { default } from './Carousel'
-export * from './Carousel'
