--- packages/ui/src/hooks/useSlider/index.ts
@@ -1,2 +1,0 @@
-export { default } from './useSlider'
-export * from './useSlider'
