+++ packages/ui/src/index.ts
@@ -12,9 +12,6 @@ export type { IncentiveProps } from './components/atoms/Incentive'
 export { default as Bullets } from './components/molecules/Bullets'
 export type { BulletsProps } from './components/molecules/Bullets'
 
-export { default as Carousel } from './components/molecules/Carousel'
-export type { CarouselProps } from './components/molecules/Carousel'
-
 export {
   default as Banner,
   BannerContent,
@@ -31,12 +28,3 @@ export type {
 // Organisms
 export { Tile, Tiles } from './components/organisms/Tiles'
 export type { TilesProps, TileProps } from './components/organisms/Tiles'
-
-// Hooks
-export { default as useSlider } from './hooks/useSlider'
-export type {
-  UseSliderArgs,
-  SliderState,
-  SliderDispatch,
-  SlideDirection,
-} from './hooks/useSlider'
